refactor: 按語言分類 E2E 測試目錄結構
tests/e2e/commands/
├── typescript/ (16 files)
├── swift/ (9 files)
└── python/ (4 files)
