+++ tests/e2e/commands/typescript/cli-cycles-coverage.e2e.test.ts
@@ -4,7 +4,7 @@
  */
 
 import { describe, it, expect, beforeEach, afterEach } from 'vitest';
-import { loadFixture, executeCLI, type FixtureContext } from '../../helpers/index.js';
+import { loadFixture, executeCLI, type FixtureContext } from '../../../helpers/index.js';
 
 describe('CLI cycles coverage - 基於 sample-project fixture', () => {
   let fixture: FixtureContext;
