+++ tests/e2e/commands/typescript/cli-move-extreme.e2e.test.ts
@@ -4,7 +4,7 @@
  */
 
 import { describe, it, expect, beforeEach, afterEach } from 'vitest';
-import { loadFixture, executeCLI, type FixtureContext } from '../../helpers/index.js';
+import { loadFixture, executeCLI, type FixtureContext } from '../../../helpers/index.js';
 import * as path from 'path';
 
 describe('CLI move extreme - 基於 sample-project fixture', () => {
