+++ tests/e2e/commands/typescript/cli-rename-coverage.e2e.test.ts
@@ -6,7 +6,7 @@
  */
 
 import { describe, it, expect, beforeEach, afterEach } from 'vitest';
-import { loadFixture, executeCLI, type FixtureContext } from '../../helpers/index.js';
+import { loadFixture, executeCLI, type FixtureContext } from '../../../helpers/index.js';
 
 describe('CLI rename coverage - 實際執行測試', () => {
   let fixture: FixtureContext;
